grpc_service/graph/pyproject.toml
[tool.poetry]
name = "graph"
version = "0.1.0"
description = ""
authors = ["Jeff Su <jeffdsu@gmail.com>"]
readme = "README.md"
packages = [{include = "grpc_try"}]

[tool.poetry.dependencies]
python = "^3.11"
grpcio = "^1.59.3"
grpcio-tools = "^1.59.3"
strawberry-graphql = {extras = ["debug-server"], version = "^0.215.1"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
